ilera-backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
Brotli==1.1.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
cssbeautifier==1.15.4
dj-database-url==2.3.0
Django==5.2.1
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djlint==1.36.4
EditorConfig==0.17.0
frozenlist==1.6.0
gunicorn==23.0.0
idna==3.10
jsbeautifier==1.15.4
json5==0.12.0
multidict==6.4.3
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
propcache==0.3.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-decouple==3.8
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requirements-parser==0.13.0
schema==0.7.7
six==1.17.0
sqlparse==0.5.3
tomli==2.2.1
tqdm==4.67.1
twilio==9.6.0
typing_extensions==4.13.2
tzdata==2025.2
ulid-py==1.1.0
urllib3==2.4.0
whitenoise==6.9.0
yarl==1.20.0
